+++ tests/bootstrap.php
@@ -18,6 +18,18 @@ if ( ! defined( 'AEOCAS_PLUGIN_FILE' ) ) {
     define( 'AEOCAS_PLUGIN_FILE', dirname( __DIR__ ) . '/aeo-content-ai-studio.php' );
 }
 
+function __( $text, $domain = null ) {
+    return $text;
+}
+
+function add_action( $hook, $callback, $priority = 10, $accepted_args = 1 ) {
+    return true;
+}
+
+function add_filter( $hook, $callback, $priority = 10, $accepted_args = 1 ) {
+    return true;
+}
+
 function update_option( $name, $value, $autoload = null ) {
     $GLOBALS['aeocas_test_options'][ $name ] = $value;
 }
@@ -27,9 +39,24 @@ function plugin_basename( $file ) {
 }
 
 $GLOBALS['aeocas_test_options'] = array();
+$GLOBALS['aeocas_test_menu_page_args'] = null;
 
 function get_option( $name, $default = false ) {
     return array_key_exists( $name, $GLOBALS['aeocas_test_options'] ) ? $GLOBALS['aeocas_test_options'][ $name ] : $default;
+}
+
+function add_menu_page( $page_title, $menu_title, $capability, $menu_slug, $callback = '', $icon_url = '', $position = null ) {
+    $GLOBALS['aeocas_test_menu_page_args'] = array(
+        'page_title' => $page_title,
+        'menu_title' => $menu_title,
+        'capability' => $capability,
+        'menu_slug'  => $menu_slug,
+        'callback'   => $callback,
+        'icon_url'   => $icon_url,
+        'position'   => $position,
+    );
+
+    return $menu_slug;
 }
 
 function site_url() {